Deletes ProtobufFieldStrategy wrapper and inlines factory logic into ProtobufStrategyExtensions
